requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.30.2
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
faker==19.6.2
xgboost==1.7.6
imbalanced-learn==0.11.0
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
Flask-Login==0.6.3
bcrypt==4.0.1
google-generativeai==0.3.2
bcrypt==4.1.3